Refactored and removed unused class
