+++ app/src/main/java/org/listenbrainz/android/model/FeedPayload.kt
@@ -2,6 +2,6 @@ package org.listenbrainz.android.model
 
 data class FeedPayload(
     val count: Int,
-    val events: List<Event>,
+    val events: List<FeedEvent>,
     val user_id: String
 )
